requirements.txt
mcp
rosbags
numpy
plotly
pyyaml